remove planButtons
